requirements.txt
altair
antlr4-python3-runtime
attrs
blinker
cachetools
certifi
charset-normalizer
click
cloudpickle
contourpy
cycler
easyocr
filelock
fonttools
fsspec
ftfy
fvcore
gitdb
GitPython
huggingface-hub
idna
imageio
iopath
Jinja2
jsonschema
jsonschema-specifications
kiwisolver
lazy_loader
markdown-it-py
MarkupSafe
matplotlib
mdurl
mpmath
networkx
ninja
numpy
omegaconf
opencv-python
opencv-python-headless
packaging
pandas
pillow
portalocker
protobuf
pyarrow
pyclipper
pycocotools
pydeck
Pygments
pyparsing
